Add test_util.cpp for api helpers
