+++ tests/backend/CMakeLists.txt
@@ -11,6 +11,7 @@ add_executable(public_hipdnn_backend_tests
     test_execution_plan_api.cpp
     test_variant_pack_api.cpp
     test_handle_api.cpp
+    test_util.cpp
 )
 
 target_compile_options(public_hipdnn_backend_tests PRIVATE ${HIPDNN_WARNING_COMPILE_OPTIONS})
